+++ lib/representable/xml/binding.rb
@@ -3,8 +3,7 @@ require 'representable/hash/binding.rb'
 
 module Representable
   module XML
-    module_function
-    def Node(document, name, attributes={})
+    module_function def Node(document, name, attributes={})
       node = Nokogiri::XML::Node.new(name.to_s, document) # Java::OrgW3cDom::DOMException: NAMESPACE_ERR: An attempt is made to create or change an object in a way which is incorrect with regard to namespaces.
 
       attributes.each { |k, v| node[k] = v } # TODO: benchmark.